WIP: Make + and - zoom in and out, respectively
This adds a zoom function to the PCanvas and implements for the two
current backends. For vispy, the zoom happens under the cursor, like how
the mouse wheel behaves. But for pygfx, I did not yet figure out how to
do that. Do we need to translate the focused point to the origin, then
change zoom, then translate it back?