refactor: Pass LUTModel to LUTView (#87)
* Convert LUTView to Model-View

* rename views

* Remove set_colormap_without_signal usage in tests

* Sync autoscale and clims

* ImageHandle.cmap -> ImageHandle.colormap

Aligns with LutView

* PyGFXImageHandle.set_gamma: Allow gamma=1

* Clean up LutView

Use Optional"[LutModel]" over "LutModel | None" for Python 3.9

Remove unused ctor

* Comment _channel_controller.py

* More cleanups

* Add tests for QLutView

* Fix qt check

* Fix pytest qt skip

:)

* Add wx to dev extra

* Add LutView tests for Jupyter, Wx

* Improve wx tests

* Move autoscale clims to controller

Since the actual value to set here is gathered from the data, not from
the user, it shouldn't really be on the view. Also will make things much
easier when multiple handles have different data.

* Fix _CmapCombo.setCurrentColormap

* Test enabling autoscaling

* Fixes

* Fix tests

* use qtbot to manage QLutView

* Test controller autoscaling logic

* Make test smaller

The use of a controller was causing failures in some places.

* Ad __eq__ for ClimPolicy impls

Will prevent a fair number of signal reemissions. Thanks to @tlambert03
for the suggestion

* Clean up jupyter lutview

Thanks again @tlambert03 for the suggestion

* Clean up qt lutview

* LutView: Strengthen method descriptions

* vispy: remove fps printout

* Cleanup

* Align src and test package names

* Add note for QLutView.set_clims

* Remove breakpoint()

* Remove unused ChannelController.clims

---------

Co-authored-by: Talley Lambert <talley.lambert@gmail.com>