feat: Better histogram control (plus pygfx histogram overhaul) (#146)
* BIG WIP: Add slider

* WIP: Log slider

* Add double click to reset gamma

* WIP: Log button instead

Not much utility to a slider

* WIP: More histogram-related niceness

* Rework pygfx histogram

Makes log work, but more importantly makes for better histogram
interaction and cleaner logic

* Add Histogram controls to wx

* Add buttons for jupyter

* qt: remove auto thresholding ignoring tails

Will save for another PR

* GraphicsCanvas: add set_extent

Useful for histograms right now but potentially for array canvases as
well?

* JupyterLutView: remove button value

Whoops

* Fix _jupyter/test_array_options

* pygfx histogram: Prevent dbz

* Remove FIXME

I don't think this applies anymore

* Compute histogram range when updating dtype

Since the goal is to not move the camera when updating data, without
this change we'll never set the camera to the right place the first
time!

* Jupyter histogram: resolve small fixmes

* Add LutModel.set_clim_bounds

Not sure how I feel about this addition to the model, but it's very
useful!

* Use Optionals for Python 3.9

* Add double-click events to jupyter, wx

* wx: Minor fixes

* Clean up double clicking

* Clean up qt changes

* vispy histogram: Remove log cruft

* plotwidget: remove commented method

* Clean up camera rect bounding logic

* Clean up wx LutView

* pygfx: margin_top <-> margin_bottom

* pygfx: clean up histogram cruft

* style(pre-commit.ci): auto fixes [...]

* Write UI unit tests

* Retain reference to histogram controls

...maybe this can resolve the macos segfaults?

* WIP: Find that segfault

* Use CallAfter to delete parent

* Test Vispy histogram interaction: first steps

* Use any_app

* Flesh out vispy histogram tests

* Add pygfx test

Still getting widget leaks...at least the test passes...

* pygfx: Allow leaks

It's not clear how to fix them...

* Remove unnecessary resize

* Try fixing widget leaks

* Set size policy

Should solve histogram taking up all of the vertical space...

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>