feat: histogram highlights (#181)
* Add a highlight to the VisPy canvas

Assists the task of finding where a pixel lies within the histogram.

* Change highlight color to yellow

* Add highlight visual to pygfx

* Add mouseLeft event

Naming could be confusing with the left mouse button...but the idea is
past tense of leave...

With the highlight there is a need to determine when the cursor leaves
the canvas, so that the highlight is hidden. The normal mouse moved
events are not sufficient if either (a) the user has panned/zoomed such
that data is visible on the edge of the canvas or (b) the user has moved
the cursor so quickly that debouncing prevents a mouse event from
occurring above the margin between the bounds of the data and the edge
of the canvas

* Implement on_mouse_leave for wx

* Test histogram highlights

Noticed during testing that vispy and pygfx differ in their handle
scaling. I like pygfx's scaling better, as the handles then were
directly scaled to be slightly taller than the data. So I changed vispy
to behave like pygfx

* Extract highlight_values helper method

Turns out we also want to hide the hover info on a mouse leave event.
Might as well just make a helper method that both moving the mouse and
leaving the canvas can both call

* Test mouse move/leave controller code

* update color

* pragma

---------

Co-authored-by: Talley Lambert <talley.lambert@gmail.com>